vehicle/requirements.txt
alembic==1.13.1
annotated-types==0.7.0
aws-lambda-powertools==2.38.1
coverage==7.5.3
exceptiongroup==1.2.1
greenlet==3.0.3
iniconfig==2.0.0
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pydantic==2.7.3
pydantic_core==2.18.4
pytest==8.2.2
pytest-mock==3.14.0
python-dotenv==1.0.1
SQLAlchemy==2.0.30
tomli==2.0.1
typing_extensions==4.12.1
